+++ src/index.ts
@@ -4,18 +4,12 @@ import { Command } from "commander";
 import { addPackage } from "./commands/add/index.js";
 import { buildSchema } from "./commands/generate/index.js";
 import { initProject } from "./commands/init/index.js";
-import { toggleAnalytics } from "./commands/init/utils.js";
 
 const program = new Command();
-program.name("kirimase").description("Kirimase CLI").version("0.1.0");
-
-program
-  .command("analytics")
-  .option("-t, --toggle", "toggle anonymous analytics")
-  .action(toggleAnalytics);
+program.name("gennext").description("GenNext CLI").version("0.1.0");
 
 addCommonOptions(program.command("init"))
-  .description("initialise and configure kirimase within directory")
+  .description("initialise and configure gennext within directory")
   .action(initProject);
 
 program